First pass at updating page contents

+++ app/models/base_model.py
@@ -5,6 +5,7 @@ import yaml
 
 class InvalidModelError(Exception):
     def __init__(self, reason, model):
+        self.reason = reason
         super().__init__("{reason}\n\tin <{classname}: {filename}>".format(reason=reason, classname=model.__class__.__name__, filename=model.filename))
 
 class BaseModel():
@@ -22,6 +23,9 @@ class BaseModel():
         self.filename = filename
         self.path = os.path.join(self.BASE_CONTENT_DIR, self.CONTENT_DIR, filename)
         self.parse_page()
+        self.__changed = False
+        self.__valid = False
+        self.errors = []
 
     def parse_page(self):
         contents = ""
@@ -34,18 +38,45 @@ class BaseModel():
         self.markdown = "---".join(contents.split("---")[2:]).strip()
         self.metadata = parsed_file.metadata
 
+    def serialize(self):
+        return {
+            "markdown": self.markdown,
+            "metadata": self.metadata,
+            "required_meta": self.REQUIRED_META,
+        }
+
     def validate(self):
+        self.errors = []
         for key in self.REQUIRED_META:
-            if key not in self.metadata:
-                raise InvalidModelError("Metadata \"{key}\" not found".format(key=key), self)
+            if key not in self.metadata or not self.metadata[key]:
+                self.errors.append("Metadata \"{key}\" is required".format(key=key))
+
+        self.__valid = self.errors == []
+        return self.__valid
 
     def save(self):
         self.validate()
 
+        if not self.__changed or not self.__valid:
+            return False
+
         with open(self.path, "r+") as f:
             f.seek(0)
-        f.write("---\n")
-        f.write(yaml.dump(self.metadata, default_flow_style=False))
-        f.write("---\n")
-        f.write(self.content)
-        f.truncate()
+            f.write("---\n")
+            f.write(yaml.dump(self.metadata, default_flow_style=False))
+            f.write("---\n\n")
+            f.write(self.content)
+            f.truncate()
+
+        return True
+
+    def update(self, **kwargs):
+        if "metadata" in kwargs:
+            if self.metadata != kwargs["metadata"]:
+                self.metadata = kwargs["metadata"]
+                self.__changed = True
+
+        if "markdown" in kwargs:
+            if self.markdown != kwargs["markdown"]:
+                self.markdown = kwargs["markdown"]
+                self.__changed = True
